Add tokenizer
<3 goto

this tokenizer sucks, mostly because I suck, but also because
 - variables are passed into the output
 - function names are passed into the output

but the github tokenizer does the same

I mean being able to tokenize every programming language in existence
is as ambitiously insane as it sounds, so whatever

I am noticing however that I didn't add something the github tokenizer
does have, which is to remove 'punctuation', i.e. braces and semicolons
I'll add that next

This is one of the few instances in which using goto makes sense,
so I'm using it.

Also even though this isn't perfect, I'm going to roll with it anyway
for the rest of this package because this is necessary, what I have now
barely qualifies as a placeholder and in the future, a drop-in replacement
for this tokenizer should be dead-easy for someone who actually knows what
they are doing.