Finish up Naive Bayesian Classification
I flailed around with this code, trying out various ideas for embedding
the data from imported package bayesian, which was the primary concern:

 - Try not relying on resolving filepaths at runtime

Messed around with embedding the data exported by bayesian.WriteTo()
into a generated go file as a base64 string. Ran into stupid issues

Eventually decided on taking advantage of the fact that go *must* have a
GOPATH or a GOROOT to make an intelligent guess as to whether the data
will be available.

 - Get the right language name from the result of calling bayesian.LogScores()

While I'm grateful that someone did the hard work of putting all that math
and machine learning stuff into a package, they sure as hell made it non-obvious
as to how to get the name of the class as a string from the result set.

That is, of course, until I tried the obvious way ~_~