Performance: read only first 512 bytes of files/blobs
Rather than using ioutil.ReadAll and exec.Command.CombinedOutput
on potentially very large files, of which only the first 512 bytes
need be read to determine if it IsBinary(), and of which every
last token needn't be parsed to be classified by the bayesian package,
(though this has yet to be proven, though preliminary results are very
promising), these changes introduce limiting the amount of data
throughput, reducing i/o latency and increasing performance
significantly on large repositories.

Note, now the latency of subshell commands, in other words calling
"git", is actually greater than using the filesystem calls for small
repos.