Experimental Idea: constrain data analysis results based on filename
I noticed that DetectFromFilename only returns an answer if there is
only one known language for a given name or extension.

If there was more than one, this information was unused.

This commit changes that.

Preliminary results are satisfactory in their increased accuracy.

Renaming/rethinking the API as a result of this change is next on the
todo list.

Please note that in the absence of language "hints", the old behavior
of relying entirely on the #1 result from bayesian.LogScores should be
invoked.