Add automatic generation to build process
Now a call to `make` ensures this can be run on any architecture/os.

There are several improvements here:

 - github.com/github/linguist is now a dependency in the form of a git
   submodule which removes the need to store the yaml files in the repo

 - the generate_classifier code has been cleaned up, commented, and made
   fully concurrent, drastically improving performance.

 - using go generate (and Makefile) to update the submodule, generate
   the classifier, and copy over the static assets (the yaml files)

Additionally, none of the generated/static data is kept in the repo
anymore, so it can be regenerated with the latest updates to the
github/linguist dataset at any time.

This is how it should have worked all along, I just wasn't sure how to
do it before.